# crl-release-22.2: db: restore TrySeekUsingNext across SetOptions batch refresh